+++ RELEASES.md
@@ -24,6 +24,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * service-templates: recursive traversal query
 * service-templates: navigation tab
 * prov-template-compiler: getting level0 types for template variables
+* service-templates: displaying level0 types colours for template variables